Renamed Scanner classes to propper style, fixed handling multi-line comments, changed output formatting, added sclex1.scl, reformatted code
